feat: tests for failed authentication requests
chore: code splitting
