+++ .gitignore
@@ -1,0 +1,2 @@
+deno.json
+development.ts
