+++ data/authentication/authenticationHeaders.ts
@@ -1,0 +1,5 @@
+export const authorizationHeaders = {
+  "Content-Type": "application/x-www-form-urlencoded",
+  "Authorization":
+    "Basic MDk1MTUxNTktNzIzNy00MzcwLTliNDAtMzgwNmU2N2MwODkxOnVjUGprYTV0bnRCMktxc1A=",
+};
