+++ src/authentication.ts
@@ -1,0 +1,1 @@
+export { authenticateWithNpsso } from "./authentication/authenticateWithNpsso.ts";
