--- src/mod.ts
@@ -1,1 +1,0 @@
-export * from "./authentication/authentication.ts";
